requirements.txt
dbt-core>=1.8
dbt-postgres>=1.8
psycopg2-binary>=2.9 